FINERACT-2180: Trigger (internal) business event when new datatable entry was added
