+++ fineract-provider/src/main/java/org/apache/fineract/infrastructure/event/business/domain/datatable/DatatableEntryUpdatedBusinessEvent.java
@@ -1,0 +1,40 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements. See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership. The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing,
+ * software distributed under the License is distributed on an
+ * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
+ * KIND, either express or implied. See the License for the
+ * specific language governing permissions and limitations
+ * under the License.
+ */
+package org.apache.fineract.infrastructure.event.business.domain.datatable;
+
+import org.apache.fineract.infrastructure.event.business.domain.NoExternalEvent;
+
+public class DatatableEntryUpdatedBusinessEvent extends DatatableEntryBusinessEvent implements NoExternalEvent {
+
+    public static final String TYPE = "DatatableEntryUpdatedBusinessEvent";
+
+    public DatatableEntryUpdatedBusinessEvent(final DatatableEntryDetails datatableEntryDetails) {
+        super(datatableEntryDetails);
+    }
+
+    @Override
+    public String getType() {
+        return TYPE;
+    }
+
+    @Override
+    public Long getAggregateRootId() {
+        return datatableEntryDetails.getEntityId();
+    }
+}
